AA updates May 2022
