+++ ExtLibs/AltitudeAngelWings.Plugin/Properties/Resources.resx
@@ -117,4 +117,129 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="AskToEnableAltitudeAngel" xml:space="preserve">
+    <value>Do you wish to enable Altitude Angel airspace management data?
+For more information visit [link;http://www.altitudeangel.com;www.altitudeangel.com]</value>
+  </data>
+  <data name="ExplicitClickPrefix" xml:space="preserve">
+    <value>Altitude Angel is currently disabled. </value>
+  </data>
+  <data name="AskToEnableCaption" xml:space="preserve">
+    <value>Altitude Angel - Enable</value>
+  </data>
+  <data name="SettingsMenuItemText" xml:space="preserve">
+    <value>Altitude Angel Settings</value>
+  </data>
+  <data name="SettingsAccountTabText" xml:space="preserve">
+    <value>Account</value>
+  </data>
+  <data name="SettingsURLDomainSuffixText" xml:space="preserve">
+    <value>URL Domain Suffix</value>
+  </data>
+  <data name="SettingsClientSecretText" xml:space="preserve">
+    <value>Client Secret</value>
+  </data>
+  <data name="SettingsClientIDText" xml:space="preserve">
+    <value>Client ID</value>
+  </data>
+  <data name="SettingsOverrideClientSettingsText" xml:space="preserve">
+    <value>Override Client Settings</value>
+  </data>
+  <data name="SettingsWhatIsThisText" xml:space="preserve">
+    <value>What is this?</value>
+  </data>
+  <data name="SettingsEnableFlightReportingText" xml:space="preserve">
+    <value>Enable Flight Reporting</value>
+  </data>
+  <data name="SettingsSignOutText" xml:space="preserve">
+    <value>Sign Out</value>
+  </data>
+  <data name="SettingsSignInText" xml:space="preserve">
+    <value>Sign In</value>
+  </data>
+  <data name="SettingsMapLayersText" xml:space="preserve">
+    <value>Map Layers</value>
+  </data>
+  <data name="SettingsPlanMapText" xml:space="preserve">
+    <value>Plan Map</value>
+  </data>
+  <data name="SettingsDataMapText" xml:space="preserve">
+    <value>Data Map</value>
+  </data>
+  <data name="SettingsDefaultLayersText" xml:space="preserve">
+    <value>Default Layers</value>
+  </data>
+  <data name="SettingsOpacityAdjustText" xml:space="preserve">
+    <value>Opacity Adjust</value>
+  </data>
+  <data name="SettingsGroundDataText" xml:space="preserve">
+    <value>Ground Data</value>
+  </data>
+  <data name="SettingsFlightReportingTabText" xml:space="preserve">
+    <value>Flight Reporting</value>
+  </data>
+  <data name="SettingsFlightReportDescriptionText" xml:space="preserve">
+    <value>Flight Report Description</value>
+  </data>
+  <data name="SettingsUseLocalConflictScopeText" xml:space="preserve">
+    <value>Use Local Conflict Scope</value>
+  </data>
+  <data name="SettingsSerialNumberText" xml:space="preserve">
+    <value>Serial Number</value>
+  </data>
+  <data name="SettingsAllowSMSContactText" xml:space="preserve">
+    <value>Allow SMS Contact</value>
+  </data>
+  <data name="SettingsICAOAddressText" xml:space="preserve">
+    <value>ICAO Address</value>
+  </data>
+  <data name="SettingsUseExistingFlightPlanIDText" xml:space="preserve">
+    <value>Use Existing Flight Plan ID</value>
+  </data>
+  <data name="SettingsCommercialFlightText" xml:space="preserve">
+    <value>Commercial Flight</value>
+  </data>
+  <data name="SettingsFlightReportDurationText" xml:space="preserve">
+    <value>Flight Report Duration (mins)</value>
+  </data>
+  <data name="SettingsContactPhoneNumberText" xml:space="preserve">
+    <value>Contact Phone Number</value>
+  </data>
+  <data name="SettingsFlightReportNameText" xml:space="preserve">
+    <value>Flight Report Name</value>
+  </data>
+  <data name="SettingsAboutTabText" xml:space="preserve">
+    <value>About</value>
+  </data>
+  <data name="SettingsDisableText" xml:space="preserve">
+    <value>Disable</value>
+  </data>
+  <data name="SettingsEnableText" xml:space="preserve">
+    <value>Enable</value>
+  </data>
+  <data name="SettingsWindowTitleText" xml:space="preserve">
+    <value>Altitude Angel Settings</value>
+  </data>
+  <data name="PluginAuthor" xml:space="preserve">
+    <value>Altitude Angel</value>
+  </data>
+  <data name="PluginName" xml:space="preserve">
+    <value>Altitude Angel</value>
+  </data>
+  <data name="MissionPlannerAdapterNotifyMessageTitle" xml:space="preserve">
+    <value>Notification</value>
+  </data>
+  <data name="MissionPlannerAdapterMessageBoxDefaultCaption" xml:space="preserve">
+    <value>Message</value>
+  </data>
+  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+  <data name="AALogo" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\AALogo.png;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </data>
+  <data name="About" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\About.html;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
+  </data>
+  <data name="MapInfoDockPanel" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\MapInfoDockPanel.html;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;utf-8</value>
+  </data>
 </root>
