+++ ExtLibs/AltitudeAngelWings.Plugin/Resources/About.html
@@ -1,0 +1,7 @@
+<html>
+<body>
+    <h1>Mission Planner</h1>
+    <h2>Altitude Angel</h2>
+    <p>These Mission Planner extensions enable use of Mission Planner with Altitude Angel's GuardianUTM platform.</p>
+</body>
+</html>
