+++ ExtLibs/AltitudeAngelWings.Plugin/Resources/MapInfoDockPanel.html
