+++ ExtLibs/AltitudeAngelWings/ApiClient/Client/FilterInfoDisplay.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace AltitudeAngelWings.ApiClient.Client
 {
