+++ ExtLibs/AltitudeAngelWings/IServiceLocatorConfiguration.cs
@@ -1,0 +1,7 @@
+namespace AltitudeAngelWings
+{
+    public interface IServiceLocatorConfiguration
+    {
+        void Configure();
+    }
+}
