+++ ExtLibs/AltitudeAngelWings/Properties/Resources.resx
@@ -118,10 +118,4 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="AALogo" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\AALogo.png;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </data>
-  <data name="MapInfoDockPanel" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\MapInfoDockPanel.html;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;utf-8</value>
-  </data>
 </root>
