+++ GCSViews/ConfigurationView/ConfigPlanner.cs
@@ -938,14 +938,6 @@ namespace MissionPlanner.GCSViews.ConfigurationView
             Settings.Instance["norcreceiver"] = chk_norcreceiver.Checked.ToString();
         }
 
-        private void but_AAsignin_Click(object sender, EventArgs e)
-        {
-#if !LIB
-            MissionPlanner.Utilities.AltitudeAngel.AltitudeAngel.Configure();
-            new Utilities.AltitudeAngel.AASettings().Show(this);
-#endif
-        }
-
         private void CMB_Layout_SelectedIndexChanged(object sender, EventArgs e)
         {
             if ((DisplayNames)CMB_Layout.SelectedItem == DisplayNames.Advanced)
