+++ GCSViews/FlightData.cs
@@ -906,14 +906,6 @@ namespace MissionPlanner.GCSViews
             POI.POIAdd(MouseDownStart);
         }
 
-        private void altitudeAngelSettingsToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-#if !LIB
-            MissionPlanner.Utilities.AltitudeAngel.AltitudeAngel.Configure();
-            new Utilities.AltitudeAngel.AASettings().Show(this);
-#endif
-        }
-
         private void BUT_abort_script_Click(object sender, EventArgs e)
         {
             scriptthread.Abort();
