+++ MissionPlanner.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31613.86
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32414.318
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MissionPlanner", "MissionPlanner.csproj", "{A2E22272-95FE-47B6-B050-9AE7E2055BF5}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -218,6 +218,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Flasher", "ExtLibs\Flasher\Flasher.csproj", "{6609C54A-71B9-4349-8E79-1DF3293BC7D0}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NMEA2000", "ExtLibs\NMEA2000\NMEA2000.csproj", "{63F10356-A09B-46B0-9AFA-FAB8326A52DD}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AltitudeAngelWings.Plugin", "ExtLibs\AltitudeAngelWings.Plugin\AltitudeAngelWings.Plugin.csproj", "{DED183CA-5591-4CE5-A89D-987035BAC553}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -535,6 +537,10 @@ Global
 		{63F10356-A09B-46B0-9AFA-FAB8326A52DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{63F10356-A09B-46B0-9AFA-FAB8326A52DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{63F10356-A09B-46B0-9AFA-FAB8326A52DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DED183CA-5591-4CE5-A89D-987035BAC553}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DED183CA-5591-4CE5-A89D-987035BAC553}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DED183CA-5591-4CE5-A89D-987035BAC553}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DED183CA-5591-4CE5-A89D-987035BAC553}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -564,7 +570,7 @@ Global
 		{BB06DFF7-4F41-4B9D-A3C3-3B6D2B8702B6} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
 		{CCE510F7-1DA6-40F2-8921-B86ED41BB85E} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
 		{65473257-E70F-410B-9269-D0C0F771EA87} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
-		{098F542C-7C70-4B9C-B2E2-AC5DC5E0294B} = {82872363-B5AC-4ACE-9200-83F5EF9AF27A}
+		{098F542C-7C70-4B9C-B2E2-AC5DC5E0294B} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
 		{CF402948-A3F4-4962-8A57-E40B46759D25} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
 		{766FCF85-4809-44DB-B316-6C4902C65245} = {3CA4F989-80BA-4814-8CC1-6D0DC1F5FF52}
 		{13D2EC90-C41F-48A1-AADA-859B6DC24EDC} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
@@ -618,6 +624,7 @@ Global
 		{FB690197-2695-489C-A47C-439FF3AEF7A8} = {9B3AC501-5FCB-4A2A-A628-BE0985DB2BD4}
 		{6609C54A-71B9-4349-8E79-1DF3293BC7D0} = {0371886E-5A61-4626-B9E0-50A837ECE2DD}
 		{63F10356-A09B-46B0-9AFA-FAB8326A52DD} = {7E264F12-9EC3-4CDC-A187-6879C2B2BFCF}
+		{DED183CA-5591-4CE5-A89D-987035BAC553} = {9B3AC501-5FCB-4A2A-A628-BE0985DB2BD4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F7AF94B5-83B0-46CD-B258-67F2404F82FC}
